[FIX]文档、评论删除权限控制优化

See merge request choerodon-agile/knowledgebase-service!43

+++ react/src/app/knowledge/components/DocComment/DocComment.js
@@ -39,9 +39,13 @@ import './DocComment.scss';
     });
   };
 
-  handleDeleteComment = (id) => {
+  handleDeleteComment = (id, mode) => {
     const { store } = this.props;
-    store.deleteComment(id);
+    if (mode === 'admin') {
+      store.adminDeleteComment(id);
+    } else {
+      store.deleteComment(id);
+    }
   };
 
   handleEditComment = (comment, data) => {
